requirements.txt
numpy==2.1.0
openai==1.43.0
playsound==1.3.0
pyaudio==0.2.14
python-dotenv==1.0.1
